Resolve cyclic import
