+++ .pre-commit-config.yaml
@@ -17,7 +17,7 @@
 # See https://pre-commit.com/hooks.html for more hooks
 repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.2.0
+    rev: v4.3.0
     hooks:
       - id: trailing-whitespace
         exclude: >
@@ -57,7 +57,7 @@ repos:
           )$
 
   - repo: https://github.com/psf/black
-    rev: 22.3.0
+    rev: 22.10.0
     hooks:
       - id: black
         exclude: >
@@ -68,7 +68,7 @@ repos:
           )$
 
   - repo: https://github.com/PyCQA/flake8
-    rev: 4.0.1
+    rev: 5.0.4
     hooks:
       - id: flake8
         exclude: >
@@ -92,14 +92,14 @@ repos:
         types_or: [python]
 
   - repo: https://github.com/PyCQA/bandit
-    rev: 1.7.1
+    rev: 1.7.4
     hooks:
       - id: bandit
         args: ["--skip=B101"]
         types_or: [python]
 
   - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: "v0.920"
+    rev: "v0.982"
     hooks:
       - id: mypy
         args: [sdv]
@@ -118,7 +118,7 @@ repos:
           )$
 
   - repo: https://github.com/Lucas-C/pre-commit-hooks
-    rev: v1.1.13
+    rev: v1.3.1
     hooks:
       - id: insert-license
         files: '.*\.(py|pyi|yaml|yml|sh|helmignore|dockerignore|gitignore)$'
