+++ sdv/vdb/types.py
@@ -1,0 +1,28 @@
+# Copyright (c) 2022 Robert Bosch GmbH and Microsoft Corporation
+#
+# This program and the accompanying materials are made available under the
+# terms of the Apache License, Version 2.0 which is available at
+# https://www.apache.org/licenses/LICENSE-2.0.
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations
+# under the License.
+#
+# SPDX-License-Identifier: Apache-2.0
+
+from typing import Generic, TypeVar
+
+from google.protobuf.timestamp_pb2 import Timestamp
+
+T = TypeVar("T")
+
+
+class TypedDataPointResult(Generic[T]):
+    """A typed data point result"""
+
+    def __init__(self, path: str, value: T, timestamp: Timestamp):
+        self.path = path
+        self.value = value
+        self.timestamp = timestamp
